Factor out the common part of the PBKDF2 algorithm into its own function. The three copies now just call that function with different PRF instances.
